tier cleanup and move lifespark
