+++ src/analysis/retail/evoker/preservation/constants.ts
@@ -88,6 +88,7 @@ export const CYCLE_OF_LIFE_SEED_DURATION = 10000;
 export const EMPATH_REGEN_FACTOR = 1;
 export const FLOW_STATE_FACTOR = 0.1;
 export const GRACE_PERIOD_INCREASE = 0.075;
+export const LIFESPARK_INCREASE = 0.5;
 export const RENEWING_BREATH_INCREASE = 0.15;
 export const TIMELESS_MAGIC = 0.15;
 export const TIMELORD_INCREASE = 0.25;
