+++ src/analysis/retail/evoker/preservation/modules/core/HotAttributor.ts
@@ -7,7 +7,7 @@ import {
   isFromDreamBreathCallOfYsera,
   isFromHardcastEcho,
   isFromTAEcho,
-} from '../../normalizers/CastLinkNormalizer';
+} from '../../normalizers/EventLinking/CastLinkNormalizer';
 import HotTrackerPrevoker from '../core/HotTrackerPrevoker';
 
 const debug = false;
