+++ src/analysis/retail/evoker/preservation/modules/talents/Echo.tsx
@@ -22,10 +22,9 @@ import {
   ECHO_TYPE,
   getEchoTypeForGoldenHour,
   getEchoTypeForLifebind,
-  isEchoFromT314PC,
   isFromHardcastEcho,
   isFromTAEcho,
-} from '../../normalizers/CastLinkNormalizer';
+} from '../../normalizers/EventLinking/CastLinkNormalizer';
 import HotTrackerPrevoker from '../core/HotTrackerPrevoker';
 
 class Echo extends Analyzer {
@@ -39,7 +38,6 @@ class Echo extends Analyzer {
   // Map<spellId, totalHealing>, only update for echo healing
   echoHealingBySpell: Map<number, number> = new Map<number, number>();
   taEchoHealingBySpell: Map<number, number> = new Map<number, number>();
-  tierEchoHealingBySpell: Map<number, number> = new Map<number, number>();
   totalApplied: number = 0;
   totalExpired: number = 0;
 
@@ -75,11 +73,7 @@ class Echo extends Analyzer {
       return;
     }
     const spellID = event.ability.guid;
-    const mapRef = this.isFromTaEcho(event)
-      ? this.taEchoHealingBySpell
-      : this.isFromHardcast(event)
-        ? this.echoHealingBySpell
-        : this.tierEchoHealingBySpell;
+    const mapRef = this.isFromTaEcho(event) ? this.taEchoHealingBySpell : this.echoHealingBySpell;
     mapRef.set(spellID, mapRef.get(spellID)! + (event.amount || 0) + (event.absorbed || 0));
   }
 
@@ -98,7 +92,7 @@ class Echo extends Analyzer {
       const hot = this.hotTracker.hots[targetID][spellID];
       return this.hotTracker.fromEchoHardcast(hot) || this.hotTracker.fromEchoTA(hot);
     }
-    return isFromHardcastEcho(event) || isFromTAEcho(event) || isEchoFromT314PC(event);
+    return isFromHardcastEcho(event) || isFromTAEcho(event);
   }
 
   isFromTaEcho(event: HealEvent) {
