+++ src/analysis/retail/evoker/preservation/modules/talents/Stasis.tsx
@@ -18,7 +18,7 @@ import STATISTIC_CATEGORY from 'parser/ui/STATISTIC_CATEGORY';
 import STATISTIC_ORDER from 'parser/ui/STATISTIC_ORDER';
 import { PassFailCheckmark, PerformanceMark, SectionHeader } from 'interface/guide';
 import { GUIDE_CORE_EXPLANATION_PERCENT } from '../../Guide';
-import { getStasisSpell, isStasisForRamp } from '../../normalizers/CastLinkNormalizer';
+import { getStasisSpell, isStasisForRamp } from '../../normalizers/EventLinking/CastLinkNormalizer';
 import { ReactNode, useState } from 'react';
 
 interface StasisInfo {
